refactor(llma): extract provider definitions into providers.rs
